Fix the renewal logic (#381)
Also look at the end time of the reservation when checking if reservation can be renewed => it can be renewed if the reservation in question is about to end within one month.

Added renewed_from column to the boat_space_reservation table. This enables us to terminate old reservation after payment has been done for the renewed reservation.