Fix flaky E2E test (#994)
* Force E2E tests to run sequentially to prevent database deadlocks

* Fix flaky E2E test

A delay in the unfinished-reservation-expiration request caused the payment
page to redirect back to the front page because the mocked current time
advanced past the timer expiration before the request could complete. This was
fixed by waiting for the timer to appear on the page before proceeding with the
test and mocking the time.

* Remove unneccessary close

* Disable async job runner when running tests

* Wait for network idle after each test to reduce the risk of concurrency-related database deadlocks