Employee can disable termination message
When terminating a boat space reservation,
an admin can choose to not include a message.

For example, an employee may be doing boat space
switches by terminating a previous and booking
a new one and there's no need to send a termination
message in those cases.

A checkbox is unchecked by default as that's the
more common use case.